Add Branch Coverage Column including refactoring common coverage column code to an abstract class
