+++ src/main/resources/hudson/plugins/jacococoveragecolumn/BranchCoverageColumn/columnHeader_de.properties
@@ -1,0 +1,2 @@
+Branch\ Coverage=Branchabdeckung
+Percentage\ of\ branches\ of\ code\ that\ are\ covered\ in\ unit\ tests=Prozentuale Branch-Testabdeckung
