+++ src/main/resources/hudson/plugins/jacococoveragecolumn/BranchCoverageColumn/column_de.properties
@@ -1,0 +1,1 @@
+Percentage\ of\ branches\ of\ code\ that\ are\ covered\ in\ unit\ tests=Prozentuale Branch-Testabdeckung
