+++ src/main/resources/hudson/plugins/jacococoveragecolumn/BranchCoverageColumn/config.jelly
@@ -1,6 +1,6 @@
 <?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
     <f:block>
-        <p>${%Shows the Java code coverage measured with JaCoCo.}</p>
+        <p>${%Shows the Java code branch coverage measured with JaCoCo.}</p>
     </f:block>
 </j:jelly>
