+++ src/main/resources/hudson/plugins/jacococoveragecolumn/BranchCoverageColumn/config_de.properties
@@ -1,0 +1,1 @@
+Shows\ the\ Java\ code\ branch\ coverage\ measured\ with\ JaCoCo.=Zeigt die Java-Code-Branchbbdeckung, die durch JaCoCo gemessen wurde.
