--- src/main/resources/hudson/plugins/jacococoveragecolumn/JaCoCoColumn/config_de.properties
@@ -1,1 +1,0 @@
-Shows\ the\ Java\ code\ coverage\ measured\ with\ JaCoCo.=Zeigt die Java-Code-Abdeckung, die durch JaCoCo gemessen wurde.
