+++ src/main/resources/hudson/plugins/jacococoveragecolumn/LineCoverageColumn/columnHeader_de.properties
