+++ src/main/resources/hudson/plugins/jacococoveragecolumn/LineCoverageColumn/column_de.properties
