+++ src/main/resources/hudson/plugins/jacococoveragecolumn/LineCoverageColumn/config.jelly
@@ -1,0 +1,6 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+    <f:block>
+        <p>${%Shows the Java code lines coverage measured with JaCoCo.}</p>
+    </f:block>
+</j:jelly>
