+++ src/main/resources/hudson/plugins/jacococoveragecolumn/LineCoverageColumn/config_de.properties
@@ -1,0 +1,1 @@
+Shows\ the\ Java\ code\ lines\ coverage\ measured\ with\ JaCoCo.=Zeigt die Java-Code-Zeilenabdeckung, die durch JaCoCo gemessen wurde.
